+++ src/Enums/Visibility.php
@@ -16,4 +16,15 @@ class Visibility extends Enum
     private const PUBLIC = 'public';
     private const PROTECTED = 'protected';
     private const PRIVATE = 'private';
+
+    public static function fromReflectionProperty(\ReflectionProperty $property): self
+    {
+        if ($property->isPublic()) {
+            return self::PUBLIC();
+        }
+        if ($property->isProtected()) {
+            return self::PROTECTED();
+        }
+        return self::PRIVATE();
+    }
 }